+++ nexus/types/src/deployment.rs
@@ -67,12 +67,12 @@ use tufaceous_artifact::ArtifactVersionError;
 
 mod blueprint_diff;
 mod blueprint_display;
-mod chicken_switches;
 mod clickhouse;
 pub mod execution;
 mod network_resources;
 mod planning_input;
 mod planning_report;
+mod reconfigurator_config;
 mod zone_type;
 
 use crate::inventory::BaseboardId;
@@ -80,16 +80,6 @@ use anyhow::anyhow;
 use anyhow::bail;
 pub use blueprint_diff::BlueprintDiffSummary;
 use blueprint_display::BpPendingMgsUpdates;
-pub use chicken_switches::PlannerConfig;
-pub use chicken_switches::PlannerConfigDiff;
-pub use chicken_switches::PlannerConfigDisplay;
-pub use chicken_switches::ReconfiguratorConfig;
-pub use chicken_switches::ReconfiguratorConfigDiff;
-pub use chicken_switches::ReconfiguratorConfigDiffDisplay;
-pub use chicken_switches::ReconfiguratorConfigDisplay;
-pub use chicken_switches::ReconfiguratorConfigParam;
-pub use chicken_switches::ReconfiguratorConfigView;
-pub use chicken_switches::ReconfiguratorConfigViewDisplay;
 pub use clickhouse::ClickhouseClusterConfig;
 use gateway_client::types::SpType;
 use gateway_types::rot::RotSlot;
@@ -141,6 +131,16 @@ pub use planning_report::PlanningZoneUpdatesStepReport;
 pub use planning_report::ZoneAddWaitingOn;
 pub use planning_report::ZoneUnsafeToShutdown;
 pub use planning_report::ZoneUpdatesWaitingOn;
+pub use reconfigurator_config::PlannerConfig;
+pub use reconfigurator_config::PlannerConfigDiff;
+pub use reconfigurator_config::PlannerConfigDisplay;
+pub use reconfigurator_config::ReconfiguratorConfig;
+pub use reconfigurator_config::ReconfiguratorConfigDiff;
+pub use reconfigurator_config::ReconfiguratorConfigDiffDisplay;
+pub use reconfigurator_config::ReconfiguratorConfigDisplay;
+pub use reconfigurator_config::ReconfiguratorConfigParam;
+pub use reconfigurator_config::ReconfiguratorConfigView;
+pub use reconfigurator_config::ReconfiguratorConfigViewDisplay;
 pub use zone_type::BlueprintZoneType;
 pub use zone_type::DurableDataset;
 pub use zone_type::blueprint_zone_type;
